+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-
+Instruction for 
+    - CTNMT (Yang et al., 2020) training
 
 
 ### Changed